Create colored-adx-di-crossings.pine
This simple script provides additional information to current ADX ( Average Directional Index ) users based on DI+ and DI- crossings (crossover/ crossunder). When DI+ is over the DI-, ADX measures the strength of the uptrend. When DI+ is under DI- the ADX measures the strength of the downtrend. This script allows you to color your ADX and the context to be more aware of this upward or downward trend.

**This script provides 4 configuration options**

* Solid or Colored ADX (red/ green)
* Colored background
* Vertical lines signaling trend reversal based on DI crossings
* Long/Short Signals based on DI crossings

I use this script on top of Squeeze Momentum to complement the understanding and also to provide additional confirmation to a primary strategy (eg.: EMA Crossings).

Happy trading and let me know in the comments if you like this script.